requirements.txt
-e git+https://github.com/abenassi/Google-Search-API.git#egg=google-search-api

apipkg==1.4
astroid==1.6.4
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.3.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
cursor==1.2.0
cycler==0.10.0
EasyProcess==0.2.3
enum34==1.1.6
execnet==1.5.0
fake-useragent==0.1.10
future==0.16.0
graphviz==0.8.3
halo==0.0.12
idna==2.6
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
log-symbols==0.0.11
lxml==4.2.1
matplotlib==2.2.2
mccabe==0.6.1
more-itertools==4.2.0
multidict==4.3.1
numpy==1.14.3
opencv-contrib-python==3.4.1.15
pep8==1.7.1
Pillow==5.1.0
pip-autoremove==0.9.1
pluggy==0.6.0
psutil==5.4.5
py==1.5.3
pycallgraph==1.0.1
pycodestyle==2.4.0
pylint==1.9.1
pyparsing==2.2.0
pyscreenshot==0.4.2
pytesseract==0.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
selenium==2.52.0
simplejson==3.15.0
six==1.11.0
snakeviz==0.4.2
spinners==0.0.19
termcolor==1.1.0
tornado==5.0.2
Unidecode==1.0.22
urllib3==1.22
vcrpy==1.12.0
vprof==0.37.4
wikipedia==1.4.0
wincertstore==0.2
wrapt==1.10.11
yappi==0.98
yarl==1.2.4
